Main Change: runCompAuton() now has 2 params
runCompAuton() now has 2 parameters.
string side -> what side are we starting on? (from before)
int autonNum -> integer value that will specify which auton we are
running.

how is autonNum incorporated in the code?

before, when we would comment out code at the competition, that code
will run only if autonNum == 1.

then, there was additional code at the end for when that part was
commented, to have the robot turn 180 to prepare for driver control,
that will only run if autonNum == 2.